refactor files and fix typo
