+++ qiita-client/.idea/dictionaries/dictionary.xml
@@ -1,0 +1,8 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="dictionary">
+    <words>
+      <w>qiita</w>
+      <w>shofucchi</w>
+    </words>
+  </dictionary>
+</component>
