+++ qiita-client/app/src/main/kotlin/io/github/shofucchi/qiita/data/api/QiitaApi.kt
@@ -1,6 +1,7 @@
-package io.github.shofucchi.qiita.data
+package io.github.shofucchi.qiita.data.api
 
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
+import io.github.shofucchi.qiita.data.model.QiitaApiModel
 import kotlinx.serialization.json.Json
 import okhttp3.MediaType.Companion.toMediaType
 import retrofit2.Retrofit
