+++ qiita-client/app/src/main/kotlin/io/github/shofucchi/qiita/data/datasource/QiitaLocalDataSource.kt
@@ -1,5 +1,10 @@
-package io.github.shofucchi.qiita.data
+package io.github.shofucchi.qiita.data.datasource
 
+import io.github.shofucchi.qiita.data.model.Article
+import io.github.shofucchi.qiita.data.dao.ArticleDao
+import io.github.shofucchi.qiita.data.model.QiitaApiModel
+import io.github.shofucchi.qiita.data.api.json
+import io.github.shofucchi.qiita.data.model.toArticle
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.withContext
