+++ qiita-client/app/src/main/kotlin/io/github/shofucchi/qiita/data/model/Article.kt
@@ -1,4 +1,4 @@
-package io.github.shofucchi.qiita.data
+package io.github.shofucchi.qiita.data.model
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
