+++ qiita-client/app/src/main/kotlin/io/github/shofucchi/qiita/data/repository/QiitaRepository.kt
@@ -1,4 +1,6 @@
-package io.github.shofucchi.qiita.data
+package io.github.shofucchi.qiita.data.repository
+
+import io.github.shofucchi.qiita.data.datasource.QiitaDataSource
 
 class QiitaRepository(private val qiitaDataSource: QiitaDataSource) {
     suspend fun fetchArticles() = this.qiitaDataSource.fetchArticles()
